+++ examples/button_counter.rb
@@ -13,7 +13,8 @@ class ButtonCounter
     window('Hello, Button!', 190, 20) {
       vertical_box {
         button {
-          text <= [self, :count, on_read: ->(count) {"Count: #{count}"}] # data-bind button text to self count, converting to string on read.
+          # data-bind button text to self count, converting to string on read.
+          text <= [self, :count, on_read: ->(count) {"Count: #{count}"}]
           
           on_clicked do
             self.count += 1